Condense down the setup process even more to 3 steps.
