+++ app/src/main/java/de/tum/in/tumcampusapp/activities/StartupActivity.java
@@ -29,7 +29,6 @@ import java.io.File;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import de.tum.in.tumcampusapp.R;
-import de.tum.in.tumcampusapp.activities.wizard.WizNavChatActivity;
 import de.tum.in.tumcampusapp.activities.wizard.WizNavExtrasActivity;
 import de.tum.in.tumcampusapp.activities.wizard.WizNavStartActivity;
 import de.tum.in.tumcampusapp.auxiliary.AccessTokenManager;
@@ -94,14 +93,8 @@ public class StartupActivity extends AppCompatActivity {
             Utils.setSetting(this, Const.BACKGROUND_MODE, true);
             Utils.setSetting(this, CardManager.SHOW_SUPPORT, true);
 
-            Intent intent;
-            if (new AccessTokenManager(this).hasValidAccessToken()) {
-                intent = new Intent(this, WizNavChatActivity.class);
-            } else {
-                intent = new Intent(this, WizNavExtrasActivity.class);
-            }
+            Intent intent = new Intent(this, WizNavExtrasActivity.class);
             intent.putExtra(Const.TOKEN_IS_SETUP, true);
-
             startActivity(intent);
             finish();
             return;
