--- app/src/main/java/de/tum/in/tumcampusapp/activities/wizard/WizNavChatActivity.java
@@ -1,226 +1,0 @@
-package de.tum.in.tumcampusapp.activities.wizard;
-
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.net.Uri;
-import android.os.Bundle;
-import android.preference.PreferenceManager;
-import android.view.View;
-import android.widget.CheckBox;
-
-import com.google.common.base.Optional;
-import com.google.gson.Gson;
-
-import java.util.List;
-
-import de.tum.in.tumcampusapp.R;
-import de.tum.in.tumcampusapp.activities.generic.ActivityForLoadingInBackground;
-import de.tum.in.tumcampusapp.auxiliary.AccessTokenManager;
-import de.tum.in.tumcampusapp.auxiliary.AuthenticationManager;
-import de.tum.in.tumcampusapp.auxiliary.Const;
-import de.tum.in.tumcampusapp.auxiliary.NetUtils;
-import de.tum.in.tumcampusapp.auxiliary.Utils;
-import de.tum.in.tumcampusapp.exceptions.NoPrivateKey;
-import de.tum.in.tumcampusapp.models.ChatMember;
-import de.tum.in.tumcampusapp.models.ChatRoom;
-import de.tum.in.tumcampusapp.models.ChatVerification;
-import de.tum.in.tumcampusapp.models.LecturesSearchRow;
-import de.tum.in.tumcampusapp.models.LecturesSearchRowSet;
-import de.tum.in.tumcampusapp.models.TUMCabeClient;
-import de.tum.in.tumcampusapp.models.managers.ChatRoomManager;
-import de.tum.in.tumcampusapp.tumonline.TUMOnlineConst;
-import de.tum.in.tumcampusapp.tumonline.TUMOnlineRequest;
-import retrofit.RetrofitError;
-
-/**
- *
- */
-public class WizNavChatActivity extends ActivityForLoadingInBackground<Void, ChatMember> {
-
-    private boolean tokenSetup;
-    private CheckBox groupChatMode, autoJoin, acceptedTerms;
-
-    public WizNavChatActivity() {
-        super(R.layout.activity_wiznav_chat);
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        overridePendingTransition(R.anim.fadein, R.anim.fadeout);
-        disableRefresh();
-
-        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-
-        // If called because app version changed remove "Step 3" and close on back pressed
-        Intent i = getIntent();
-        if (i != null && i.hasExtra(Const.TOKEN_IS_SETUP)) {
-            tokenSetup = i.getBooleanExtra(Const.TOKEN_IS_SETUP, false);
-        }
-
-        // Get handles to all UI elements
-        groupChatMode = (CheckBox) findViewById(R.id.chk_group_chat);
-        autoJoin = (CheckBox) findViewById(R.id.chk_auto_join_chat);
-        acceptedTerms = (CheckBox) findViewById(R.id.chk_group_chat_terms);
-
-        if (new AccessTokenManager(this).hasValidAccessToken()) {
-            groupChatMode.setChecked(preferences.getBoolean(Const.GROUP_CHAT_ENABLED, true));
-            autoJoin.setChecked(preferences.getBoolean(Const.AUTO_JOIN_NEW_ROOMS, true));
-            acceptedTerms.setChecked(preferences.getBoolean(Const.GROUP_CHAT_ENABLED, true));
-        } else {
-            groupChatMode.setChecked(false);
-            groupChatMode.setEnabled(false);
-            autoJoin.setEnabled(false);
-            acceptedTerms.setEnabled(false);
-        }
-    }
-
-    public void onClickTerms(View view) {
-        Uri uri = Uri.parse(Const.CHATTERMS_URL);
-        Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-        startActivity(intent);
-    }
-
-    /**
-     * If back key is pressed, open previous activity
-     */
-    @Override
-    public void onBackPressed() {
-        finish();
-        if (!tokenSetup) {
-            if (new AccessTokenManager(this).hasValidAccessToken()) {
-                startActivity(new Intent(this, WizNavCheckTokenActivity.class));
-            } else {
-                startActivity(new Intent(this, WizNavStartActivity.class));
-            }
-            overridePendingTransition(R.anim.fadein, R.anim.fadeout);
-        }
-    }
-
-    /**
-     * Open next activity on skip
-     *
-     * @param skip Skip button handle
-     */
-    @SuppressWarnings("UnusedParameters")
-    public void onClickSkip(View skip) {
-        startNextActivity(true);
-    }
-
-    /**
-     * If next is pressed, check if token has been activated
-     *
-     * @param next Next button handle
-     */
-    @SuppressWarnings("UnusedParameters")
-    public void onClickNext(View next) {
-        if (groupChatMode.isChecked() && !acceptedTerms.isChecked()) {
-            Utils.showToast(this, R.string.have_to_accept_terms);
-            return;
-        }
-
-        startLoading();
-    }
-
-    /**
-     * Opens next wizard page
-     */
-    private void startNextActivity(boolean skip) {
-        //Stop the current
-        finish();
-
-
-        Intent intent;
-        if (skip) {
-            intent = new Intent(this, WizNavExtrasActivity.class);
-        } else {
-            intent = new Intent(this, WizNavExtrasActivity.class);
-        }
-        intent.putExtra(Const.TOKEN_IS_SETUP, tokenSetup);
-        startActivity(intent);
-        overridePendingTransition(R.anim.fadein, R.anim.fadeout);
-    }
-
-    @Override
-    protected ChatMember onLoadInBackground(Void... arg) {
-        if (groupChatMode.isChecked()) {
-            if (!NetUtils.isConnected(this)) {
-                showNoInternetLayout();
-                return null;
-            }
-
-            // Get all of the users lectures and save them as possible chat rooms
-            // TODO why do we need to do this here? Is this not done inside the updating of the cache?
-            ChatRoomManager manager = new ChatRoomManager(this);
-            TUMOnlineRequest<LecturesSearchRowSet> requestHandler = new TUMOnlineRequest<>(TUMOnlineConst.LECTURES_PERSONAL, this, true);
-            Optional<LecturesSearchRowSet> lecturesList = requestHandler.fetch();
-            if (lecturesList.isPresent() && lecturesList.get().getLehrveranstaltungen() != null) {
-                List<LecturesSearchRow> lectures = lecturesList.get().getLehrveranstaltungen();
-                manager.replaceInto(lectures);
-            } else {
-                Utils.showToastOnUIThread(this, R.string.no_rights_to_access_lectures);
-                return null;
-            }
-
-            // Get the users full name
-            String id = Utils.getSetting(this, Const.CHAT_ROOM_DISPLAY_NAME, "");
-
-            // Get the users lrzId and initialise chat member
-            String lrzId = Utils.getSetting(this, Const.LRZ_ID, "");
-            ChatMember currentChatMember = new ChatMember(lrzId);
-            currentChatMember.setDisplayName(id);
-
-            // Tell the server the new member
-            ChatMember member;
-            try {
-                // After the user has entered their display name, send a request to the server to create the new member
-                member = TUMCabeClient.getInstance(this).createMember(currentChatMember);
-            } catch (RetrofitError e) {
-                Utils.log(e);
-                Utils.showToastOnUIThread(this, R.string.error_setup_chat_member);
-                return null;
-            }
-
-            //Catch a possible error, when we didn't get something returned
-            if (member == null || member.getLrzId() == null) {
-                Utils.showToastOnUIThread(this, R.string.error_setup_chat_member);
-                return null;
-            }
-
-            // Generate the private key and upload the public key to the server
-            AuthenticationManager am = new AuthenticationManager(this);
-            if (!am.generatePrivateKey(member)) {
-                Utils.showToastOnUIThread(this, getString(R.string.failure_uploading_public_key));
-            }
-
-            // Try to restore already joined chat rooms from server
-            try {
-                List<ChatRoom> rooms = TUMCabeClient.getInstance(this).getMemberRooms(member.getId(), new ChatVerification(this, member));
-                new ChatRoomManager(this).replaceIntoRooms(rooms);
-
-                //Store that this key was activated
-                Utils.setInternalSetting(this, Const.PRIVATE_KEY_ACTIVE, true);
-
-                return member;
-            } catch (RetrofitError e) {
-                Utils.log(e);
-            } catch (NoPrivateKey e) {
-                Utils.log(e);
-            }
-        }
-        return null;
-    }
-
-    @Override
-    protected void onLoadFinished(ChatMember member) {
-        if (member != null) {
-            Utils.setSetting(this, Const.GROUP_CHAT_ENABLED, groupChatMode.isChecked());
-            Utils.setSetting(this, Const.AUTO_JOIN_NEW_ROOMS, groupChatMode.isChecked() && autoJoin.isChecked());
-            Utils.setSetting(this, Const.CHAT_MEMBER, member);
-            Utils.log("Set member to settings: " + new Gson().toJson(member));
-            startNextActivity(false);
-        } else {
-            showLoadingEnded();
-        }
-    }
-}
