--- app/src/main/res/layout/activity_wiznav_activate_key.xml
@@ -1,127 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <!-- The toolbar aka SupportActionBar -->
-    <include layout="@layout/toolbar" />
-
-    <include
-        layout="@layout/layout_all_errors"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent" />
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-
-        <TextView
-            android:id="@+id/pager_title_strip"
-            style="@style/SectionHeadline"
-            android:text="@string/startup_wizard_uppercase" />
-
-        <android.support.v4.widget.SwipeRefreshLayout
-            android:id="@+id/ptr_layout"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_alignBottom="@id/divider"
-            android:layout_below="@id/pager_title_strip">
-
-            <ScrollView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="vertical"
-                    android:padding="@dimen/padding_default">
-
-                    <TextView
-                        style="?android:attr/listSeparatorTextViewStyle"
-                        android:layout_height="wrap_content"
-                        android:text="@string/step_4"
-                        android:textColor="@color/text_light_gray" />
-
-                    <TextView
-                        android:id="@+id/tvSucc"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:padding="@dimen/padding_default" />
-
-                    <TextView
-                        android:id="@+id/tvBrowse"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:padding="@dimen/padding_default" />
-
-                    <ImageView
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:paddingBottom="@dimen/padding_default"
-                        android:paddingTop="@dimen/padding_default"
-                        android:src="@drawable/key1" />
-
-                    <ImageView
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:paddingBottom="@dimen/padding_default"
-                        android:paddingTop="@dimen/padding_default"
-                        android:src="@drawable/key2" />
-
-                    <ImageView
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:paddingBottom="@dimen/padding_default"
-                        android:paddingTop="@dimen/padding_default"
-                        android:src="@drawable/key3" />
-
-                </LinearLayout>
-            </ScrollView>
-
-        </android.support.v4.widget.SwipeRefreshLayout>
-
-        <View
-            android:id="@+id/divider"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:layout_above="@+id/buttonBar"
-            android:background="?android:attr/listDivider" />
-
-        <LinearLayout
-            android:id="@+id/buttonBar"
-            style="@style/ButtonBar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:orientation="horizontal">
-
-            <Button
-                style="@style/ButtonBarButton"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?attr/selectableItemBackground"
-                android:onClick="onClickSkip"
-                android:text="@string/button_skip" />
-
-            <View
-                android:layout_width="1dp"
-                android:layout_height="match_parent"
-                android:background="?android:attr/listDivider" />
-
-            <Button
-                style="@style/ButtonBarButton"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?attr/selectableItemBackground"
-                android:onClick="onClickNext"
-                android:text="@string/button_next" />
-        </LinearLayout>
-    </RelativeLayout>
-</LinearLayout>
