--- app/src/main/res/layout/activity_wiznav_chat.xml
@@ -1,139 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <!-- The toolbar aka SupportActionBar -->
-    <include layout="@layout/toolbar" />
-
-
-    <include
-        layout="@layout/layout_all_errors"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent" />
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:id="@+id/pager_title_strip"
-            style="@style/SectionHeadline"
-            android:text="@string/startup_wizard_uppercase" />
-
-        <android.support.v4.widget.SwipeRefreshLayout
-            android:id="@+id/ptr_layout"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_alignBottom="@id/divider"
-            android:layout_below="@id/pager_title_strip">
-
-            <ScrollView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_above="@id/divider"
-                android:layout_alignParentEnd="true"
-                android:layout_alignParentLeft="true"
-                android:layout_alignParentRight="true"
-                android:layout_alignParentStart="true"
-                android:layout_below="@id/pager_title_strip">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="vertical"
-                    android:padding="@dimen/padding_default">
-
-                    <TextView
-                        style="?android:attr/listSeparatorTextViewStyle"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:text="@string/step_3"
-                        android:textColor="@color/text_light_gray" />
-
-                    <CheckBox
-                        android:id="@+id/chk_group_chat"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="16dp"
-                        android:text="@string/lecture_chat"
-                        android:textStyle="bold" />
-
-                    <TextView
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:padding="@dimen/padding_default"
-                        android:text="@string/set_lecture_chat" />
-
-                    <CheckBox
-                        android:id="@+id/chk_auto_join_chat"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="16dp"
-                        android:text="@string/auto_join_chat"
-                        android:textStyle="bold" />
-
-                    <TextView
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:padding="@dimen/padding_default"
-                        android:text="@string/set_auto_join_chat" />
-
-                    <CheckBox
-                        android:id="@+id/chk_group_chat_terms"
-                        android:layout_width="wrap_content"
-                        android:layout_height="fill_parent"
-                        android:text="@string/accept_chat_terms" />
-
-                    <Button
-                        style="@style/FlatButton"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:onClick="onClickTerms"
-                        android:text="@string/open_chat_terms" />
-
-                </LinearLayout>
-            </ScrollView>
-        </android.support.v4.widget.SwipeRefreshLayout>
-
-        <View
-            android:id="@+id/divider"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:layout_above="@+id/buttonBar"
-            android:background="?android:attr/listDivider" />
-
-        <LinearLayout
-            android:id="@+id/buttonBar"
-            style="@style/ButtonBar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:orientation="horizontal">
-
-            <Button
-                style="@style/ButtonBarButton"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?attr/selectableItemBackground"
-                android:onClick="onClickSkip"
-                android:text="@string/button_skip" />
-
-            <View
-                android:layout_width="1dp"
-                android:layout_height="match_parent"
-                android:background="?android:attr/listDivider" />
-
-            <Button
-                style="@style/ButtonBarButton"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:background="?attr/selectableItemBackground"
-                android:onClick="onClickNext"
-                android:text="@string/button_next" />
-        </LinearLayout>
-
-    </RelativeLayout>
-</LinearLayout>
